refactor(mpi): change folder structure
