--- core_v2/mpi/interfaces/Financiador.interface.ts
@@ -1,13 +1,0 @@
-import { Schema, Document } from 'mongoose';
-
-export interface IFinanciador {
-    codigoPuco: Number;
-    nombre: String;
-    financiador: String;
-    // id: Schema.Types.ObjectId;
-    numeroAfiliado: String;
-    prepaga: Boolean;
-    idObraSocial: Number;
-}
-
-export interface IFinanciadorDoc extends Document, IFinanciador {}
