--- core_v2/mpi/interfaces/Parentesco.interface.ts
@@ -1,8 +1,0 @@
-import { Document } from 'mongoose';
-
-export interface IParentesco {
-    nombre: String;
-    opuesto: String;
-}
-
-export interface IParentescoDoc extends Document, IParentesco {}
