+++ core_v2/mpi/pacientes/paciente.controller.ts
@@ -1,6 +1,6 @@
 
 
-import { Paciente } from './paciente.schema';
+import { Paciente, PacienteMpi } from './paciente.schema';
 import * as mongoose from 'mongoose';
 import * as express from 'express';
 import * as moment from 'moment';
@@ -8,214 +8,180 @@ import { Auth } from '../../../auth/auth.class';
 import { PacienteTx } from './pacienteTx';
 import { log } from '@andes/log';
 import { logKeys } from '../../../config';
-import { EventCoreV2 } from '@andes/event-bus';
+import { EventCore } from '@andes/event-bus';
 import { IPaciente, IPacienteDoc } from './paciente.interface';
 import { Matching } from '@andes/match';
 import * as config from '../../../config';
-import { MongoQuery, isSelected } from '@andes/query-builder';
-import { getObraSocial } from '../../../modules/obraSocial/controller/obraSocial';
-
-/**
- * Crea un objeto paciente
- */
-
-export function make(body: IPaciente) {
-    const paciente = new Paciente();
-    paciente.set(body);
-    return paciente;
-}
+import {parseStr, rangoFechas} from './../../../shared/queryBuilder';
 
 
 /**
- * Guarda un objecto paciente en mongo y elastic.
- * Sincroniza con ElasticSearch si es necesario.
+ * Persiste un nuevo paciente en la base de datos ANDES y los sincroniza con ElasticSearch.
  *
- * @param {IPacienteDoc} body Datos del paciente
+ * @param {IPaciente} body Datos del paciente
  * @param {express.Request} req Request de Express para obtener los datos del usuario
  */
 
-export async function store(paciente: IPacienteDoc, req: express.Request, events = true) {
-    const session = await Paciente.db.startSession();
-    session.startTransaction();
-    const isNew = paciente.isNew;
-    try {
-        const pacienteOriginal = paciente.original();
-        const pacienteFields = paciente.modifiedPaths();
 
+export async function createPaciente(body: IPaciente, req) {
+    const session = await Paciente.db.startSession();
+    try {
+        session.startTransaction();
+        const paciente = new Paciente();
+        paciente.set(body);
         Auth.audit(paciente, req);
-        await paciente.save({ session });
-        if (isNew || paciente.sincroniza()) {
-            await PacienteTx.sync(paciente);
-        }
-        if (isNew) {
-            log(req, logKeys.mpiInsert.key, paciente._id, logKeys.mpiInsert.operacion, paciente, null);
-            await session.commitTransaction();
-            if (events) {
-                EventCoreV2.emitAsync('mpi:patient:create', paciente);
-            }
-        } else {
-            log(req, logKeys.mpiUpdate.key, paciente._id, logKeys.mpiUpdate.operacion, paciente, pacienteOriginal);
-            await session.commitTransaction();
-            if (events) {
-                EventCoreV2.emitAsync('mpi:patient:update', paciente, pacienteFields);
-            }
-        }
+        await paciente.save({ session});
+        await PacienteTx.create(paciente);
+        log(req, logKeys.mpiInsert.key, paciente._id, logKeys.mpiInsert.operacion, paciente, null);
+        await session.commitTransaction();
+        EventCore.emitAsync('mpi:patient:create', paciente);
         return paciente;
     } catch (error) {
-        if (isNew) {
-            log(req, logKeys.mpiInsert.key, null, logKeys.mpiInsert.operacion, paciente, 'Error insertando paciente');
-        } else {
-            log(req, logKeys.mpiUpdate.key, paciente._id, logKeys.mpiUpdate.operacion, null, 'Error actualizando paciente');
-        }
-        await session.abortTransaction();
-        if (error.name === 'ValidationError') {
-            error.status = 400;
-        }
+        log(req, logKeys.mpiInsert.key, null, logKeys.mpiInsert.operacion, body, 'Error insertando paciente');
+        session.abortTransaction();
         throw error;
     }
 }
 
 /**
- * Modifica los campos de un paciente.
- * @param {IPacienteDoc} paciente paciente a modifciar
- * @param {object} body Datos a modificar del paciente
+ * Actualiza un paciente existente. Si esta en MPI, lo crea en ANDES. Sino lo actualiza.
+ * Sincroniza con ElasticSearch si es necesario.
+ *
+ * @param {IPaciente} body Datos del paciente
+ * @param {express.Request} req Request de Express para obtener los datos del usuario
  */
 
-export function set(paciente: IPacienteDoc, body: any) {
-    if (paciente.estado === 'validado') {
-        delete body['documento'];
-        delete body['apellido'];
-        delete body['nombre'];
-        delete body['sexo'];
-        delete body['fechaNacimiento'];
-        delete body['estado'];
+export async function updatePaciente(paciente: IPacienteDoc, req: express.Request) {
+    const session = await Paciente.db.startSession();
+    session.startTransaction();
+    try {
+        const pacienteOriginal = paciente.toObject();
+        const pacienteFields = paciente.modifiedPaths();
+        Auth.audit(paciente, req);
+        await paciente.save({ session });
+        if (paciente.sincroniza()) {
+            await PacienteTx.sync(paciente);
+        }
+        EventCore.emitAsync('mpi:patient:update', paciente, pacienteFields);
+        log(req, logKeys.mpiUpdate.key, paciente._id, logKeys.mpiUpdate.operacion, paciente, pacienteOriginal);
+        await session.commitTransaction();
+        return paciente;
+
+    } catch (error) {
+        log(req, logKeys.mpiUpdate.key, req.body._id, logKeys.mpiUpdate.operacion, null, 'Error actualizando paciente');
+        session.abortTransaction();
+        throw error;
     }
-    paciente.set(body);
-    return paciente;
 }
+
+type DatabaseType = 'andes' | 'mpi';
+
+/**
+ * @typedef {Promise} IFindById
+ * @prop {String} db - Base de datos donde pertenece el paciente
+ * @prop {String} paciente - Objecto paciente
+ */
+
+type IFindById = Promise<{
+    db: DatabaseType,
+    paciente: IPacienteDoc
+}>;
 
 /**
  * Busca un paciente por ID. Tanto en ANDES y MPI.
  * @param {string} id ID del paciente a buscar.
  * @param {object} options
  * @param {string} options.fields Listado de campos para projectar
- * @returns {IPacienteDoc}
+ * @returns {IFindById}
  */
 
-export async function findById(id: string | String | mongoose.Types.ObjectId, options = null): Promise<IPacienteDoc> {
+export async function findById(id: string | String | mongoose.Types.ObjectId, options = null): IFindById {
+    function makeFindById(Schema, select) {
+        const queryFind = Schema.findById(id);
+        if (select) {
+            queryFind.select(select);
+        }
+        return queryFind;
+    }
     options = options || {};
     const { fields } = options;
-
-    const queryFind = Paciente.findById(id);
-    if (fields) {
-        queryFind.select(fields);
-    }
-    const paciente = await queryFind;
+    let paciente = await makeFindById(Paciente, fields);
     if (paciente) {
-        if (isSelected(fields, 'financiador')) {
-            paciente.financiador = await getObraSocial(paciente);
+        return {
+            db: 'andes',
+            paciente
+        };
+    } else {
+        paciente = await makeFindById(PacienteMpi, fields);
+        if (paciente) {
+            return {
+                db: 'mpi',
+                paciente
+            };
         }
-        return paciente;
     }
     return null;
 }
 
 /**
+ * Search de paciente en ElasticSearch
  *
+ * @param {String} type Tipo de busqueda. Enum: 'simplequery', 'multimatch', 'suggest'
+ * @param {Object} query Condiciones a buscar
  */
-export async function remove(paciente: IPacienteDoc, req: express.Request) {
-    const session = await Paciente.db.startSession();
-    session.startTransaction();
-    try {
-        paciente.$session(session);
-        await paciente.remove();
-        await PacienteTx.delete(paciente);
-        await session.commitTransaction();
-        log(req, logKeys.mpiDelete.key, paciente._id, logKeys.mpiDelete.operacion, paciente, null);
-        EventCoreV2.emitAsync('mpi:patient:delete', paciente);
-        return;
-    } catch (err) {
-        await session.abortTransaction();
-        throw err;
-    }
-}
-
-/**
- * Busca pacientes dada de una cadena de texto por documento, apellido o nombre
- *
- * @param {string} query Condiciones a buscar
- */
-export async function search(searchText: string) {
-    let body = {
-        bool: {
-            must: {
-                multi_match: {
-                    query: searchText,
-                    type: 'cross_fields',
-                    fields: [
-                        'apellido^5',
-                        'nombre^5',
-                        'alias^5',
-                        'documento^5',
-                        'apellido.ngram',
-                        'nombre.ngram',
-                        'documento.ngram',
-                        'alias.ngram'],
-                    operator: 'and'
+export async function search(type: 'simplequery' | 'multimatch' | 'suggest', query: any) {
+    let body;
+    switch (type) {
+        case 'simplequery':
+            body = {
+                simple_query_string: {
+                    query: `"${query.documento}" +"${query.apellido}" +"${query.nombre}" +"${query.sexo}"`,
+                    fields: ['documento', 'apellido', 'nombre', 'sexo'],
+                    default_operator: 'and'
                 }
-            },
-            filter: {
-                term: { activo: 'true' }
-            }
-        }
-    };
+            };
+
+            break;
+        case 'multimatch':
+            body = {
+                bool: {
+                    must: {
+                        multi_match: {
+                            query: query.cadenaInput,
+                            type: 'cross_fields',
+                            fields: ['documento', 'apellido^5', 'nombre^4'],
+                            operator: 'and'
+                        }
+                    },
+                    filter: {
+                        term: { activo: 'true' }
+                    }
+                }
+            };
+
+            break;
+        case 'suggest':
+            body = {
+                bool: {
+                    must: {
+                        match: {
+                            documento: {
+                                query: query.documento,
+                                minimum_should_match: 3,
+                                fuzziness: 2
+                            }
+                        }
+                    },
+                    filter: {
+                        term: { activo: 'true' }
+                    }
+                }
+            };
+            break;
+    }
     const pacientes = await PacienteTx.search({ query: body });
     return pacientes;
-}
 
-/**
- * Busca paciente similares a partir de su documento
- * Devuelve una lista de pacientes ordenada por score de forma descendente
- *
- * @param {object} query Condiciones a buscar
- *
- * [TODO] Definir el tipado de esta funcion
- */
-
-export async function suggest(query: any) {
-    if (query && query.documento) {
-        const body = {
-            bool: {
-                must: {
-                    match: {
-                        documento: {
-                            query: query.documento,
-                            minimum_should_match: 3,
-                            fuzziness: 2
-                        }
-                    }
-                },
-                filter: {
-                    term: { activo: 'true' }
-                }
-            }
-        };
-        const sortScore = (a, b) => {
-            return b._score - a._score;
-        };
-        let pacientes = await PacienteTx.search({ query: body });
-        if (query.id) {
-            pacientes = pacientes.filter(p => (p.id !== query.id));
-        }
-        pacientes.forEach((paciente) => {
-            const value = matching(paciente, query);
-            paciente._score = value;
-        });
-
-        return pacientes.sort(sortScore);
-    } else {
-        return [];
-    }
 }
 
 /**
@@ -229,7 +195,7 @@ export async function suggest(query: any) {
  */
 
 export function matching(pacienteA: IPaciente | IPacienteDoc, pacienteB: IPaciente, weightsDefault?: any) {
-    const personaA = {
+    let personaA = {
         documento: pacienteA.documento ? pacienteA.documento.toString() : '',
         nombre: pacienteA.nombre ? pacienteA.nombre : '',
         apellido: pacienteA.apellido ? pacienteA.apellido : '',
@@ -237,7 +203,7 @@ export function matching(pacienteA: IPaciente | IPacienteDoc, pacienteB: IPacien
         sexo: pacienteA.sexo ? pacienteA.sexo : ''
     };
 
-    const personaB = {
+    let personaB = {
         documento: pacienteB.documento ? pacienteB.documento.toString() : '',
         nombre: pacienteB.nombre ? pacienteB.nombre : '',
         apellido: pacienteB.apellido ? pacienteB.apellido : '',
@@ -245,92 +211,105 @@ export function matching(pacienteA: IPaciente | IPacienteDoc, pacienteB: IPacien
         sexo: pacienteB.sexo ? pacienteB.sexo : ''
     };
 
-    const match = new Matching();
+    let match = new Matching();
 
-    const valorMatching = match.matchPersonas(personaA, personaB, weightsDefault ? weightsDefault : config.mpi.weightsDefault, config.algoritmo);
+    let valorMatching = match.matchPersonas(personaA, personaB, weightsDefault ? weightsDefault : config.mpi.weightsDefault , config.algoritmo);
     return valorMatching;
 
 }
 
-
-const searchSpecification = {
-    identificadores: (value, query) => {
-        return MongoQuery.queryArray('identificadores', value, 'entidad', 'valor');
-    },
-    relaciones: (value, query) => {
-        return MongoQuery.queryArray('relaciones', value, 'relacion.nombre', 'referencia');
-    },
-    documento: MongoQuery.partialString,
-    nombre: MongoQuery.partialString,
-    apellido: MongoQuery.partialString,
-    fechaNacimiento: MongoQuery.matchDate,
-    sexo: MongoQuery.matchString,
-    estado: MongoQuery.matchString,
-    activo: MongoQuery.matchString,
-    barrio: {
-        filed: 'direccion.ubicacion.barrio.nombre',
-        fn: MongoQuery.partialString
-    },
-    localidad: {
-        filed: 'direccion.ubicacion.localidad.nombre',
-        fn: MongoQuery.partialString
-    },
-    provincia: {
-        filed: 'direccion.ubicacion.provincia.nombre',
-        fn: MongoQuery.partialString
-    },
-    pais: {
-        filed: 'direccion.ubicacion.pais.nombre',
-        fn: MongoQuery.partialString
-    },
-    nacionalidad: MongoQuery.partialString,
-    email: (value) => {
-        return MongoQuery.queryArray('contacto', [`email|${value}`], 'tipo', 'valor');
-    },
-    celular: (value) => {
-        return MongoQuery.queryArray('contacto', [`celular|${value}`], 'tipo', 'valor');
-    },
-    fijo: (value) => {
-        return MongoQuery.queryArray('contacto', [`fijo|${value}`], 'tipo', 'valor');
-    }
-};
-
 /**
  * Realiza una búsqueda de pacientes dada una condición
- * @param condicion
- * @param fields Selecciona los campos de los documentos a devolver
- * @returns Devuelve listado de paciente encontrados
+ *@param condicion
+ *@param fields Setea los campos de los documentos a devolver
+ *@returns
  */
-export async function find(condicion, options?: any) {
-    options = options || {};
-    const { fields, skip, limit } = options;
+export async function findPaciente(condicion, fields?: string) {
 
-    const opciones = MongoQuery.buildQuery(condicion, searchSpecification);
+    try {
 
-    let pacientesQuery = Paciente.find(opciones);
+        const opciones = {};
+        // identificadores [{'Entidad1'| valor}... {'EntidadN'| valorN} ]
+        if (condicion.indentificadores) {
+            let conds = [];
+           // verifica los identificadores
+            condicion.indentificadores.forEach(identificador => {
+                let ids = identificador.split('|');
+                let filtro;
+                if (ids[0]) {
+                    filtro['identificadores.entidad'] = ids[0];
+                }
+                if (ids[1]) {
+                    filtro['identificadores.valor'] = ids[1];
+                }
+                conds.push(filtro);
 
-    if (fields) {
-        pacientesQuery.select(fields);
+            });
+            opciones['identificadores'] = {$elemMatch: conds
+            };
+        }
+
+        if (condicion.documento) {
+            opciones['documento'] =  parseStr(condicion.documento);
+        }
+        if (condicion.nombre) {
+            opciones['nombre'] =  parseStr(condicion.nombre);
+        }
+        if (condicion.apellido) {
+            opciones['apellido'] = parseStr(condicion.apellido);
+        }
+        if (condicion.fechaNacimiento) {
+            opciones['fechaNacimiento'] = rangoFechas(condicion.fechaNacimiento);
+        }
+        if (condicion.estado) {
+            opciones['estado'] = condicion.estado;
+        }
+        if (condicion.activo) {
+            opciones['activo'] = condicion.activo;
+        }
+        if (condicion.localidad) {
+            opciones['direccion.ubicacion.localidad.nombre'] = parseStr(condicion.localidad);
+        }
+        if (condicion.barrio) {
+            opciones['direccion.ubicacion.barrio.nombre'] =  parseStr(condicion.barrio);
+        }
+        if (condicion.provincia) {
+            opciones['direccion.ubicacion.provincia.nombre'] = parseStr(condicion.provincia);
+        }
+        if (condicion.pais) {
+            opciones['direccion.ubicacion.pais.nombre'] = parseStr(condicion.pais);
+        }
+        if (condicion.nacionalidad) {
+            opciones['nacionalidad'] = parseStr(condicion.nacionalidad);
+        }
+
+        let contactos = [];
+        if (condicion.email) {
+            contactos.push({tipo: 'email', valor: parseStr(condicion.email)});
+        }
+        if (condicion.celular) {
+            contactos.push({tipo: 'celular', valor: parseStr(condicion.celular)});
+        }
+        if (condicion.fijo) {
+            contactos.push({tipo: 'fijo', valor: parseStr(condicion.fijo)});
+        }
+        if (contactos.length) {
+            opciones['contactos'] = {$elemMatch: contactos};
+        }
+
+        let pacientesAndes = await Paciente.find(opciones).select(fields);
+        let pacientesMPI = await PacienteMpi.find(opciones).select(fields);
+        let pacientes = [];
+        pacientesAndes.forEach(p => {
+            if ((pacientesMPI.findIndex(pMpi => pMpi.id === p.id)) === -1) {
+                pacientes.push(p);
+            }
+        });
+        pacientes.concat(pacientesMPI);
+
+        return pacientes;
+    } catch (error) {
+        throw error;
     }
-    if (limit) {
-        pacientesQuery.limit(limit);
-    }
-    if (skip) {
-        pacientesQuery.skip(skip);
-    }
-
-    return await pacientesQuery;
 
 }
-
-export const PacienteCtr = {
-    make,
-    store,
-    set,
-    remove,
-    search,
-    matching,
-    suggest,
-    find,
-    findById
-};
